+++ src/vpsmonitor/collectors/__init__.py
@@ -1,6 +1,7 @@
 from .aliyun_swas import AliyunSwasCollector
 from .bandwagon import BandwagonCollector
 from .panstar import PanstarCollector
+from .solusvm import RackNerdCollector
 from .virtfusion import VirtFusionCollector
 from .virtualizor import VirtualizorCollector
 
@@ -8,6 +9,7 @@ __all__ = [
     "AliyunSwasCollector",
     "BandwagonCollector",
     "PanstarCollector",
+    "RackNerdCollector",
     "VirtFusionCollector",
     "VirtualizorCollector",
 ]